+++ lib/src/domain/push_data_models/buttons_data/value_objects/button_data_continue_after_action_value.dart
@@ -1,0 +1,21 @@
+import 'package:value_object_pattern/value_object.dart';
+
+class ButtonDataContinueAfterActionValue extends ValueObject<bool> {
+  ButtonDataContinueAfterActionValue({
+    super.defaultValue = false,
+    super.isRequired = false,
+  });
+
+  @override
+  bool doParse(String? parseValue) {
+    switch (parseValue) {
+      case '1':
+      case 'true':
+      case 'True':
+      case 'TRUE':
+        return true;
+      default:
+        return false;
+    }
+  }
+}