+++ lib/src/presentation/modal_bottom_sheet/modal_bottom_sheet_content.dart
@@ -14,6 +14,7 @@ class PushModalBottomSheetContent extends PushWidget {
     super.gatekeeper,
     super.optionsBuilder,
     super.onStepSubmit,
+    super.stepValidator,
     super.onGateBlocked,
   });