+++ lib/src/presentation/snackbar/push_snack_bar_content.dart
@@ -14,6 +14,7 @@ class PushSnackBarContent extends PushWidget {
     super.gatekeeper,
     super.optionsBuilder,
     super.onStepSubmit,
+    super.stepValidator,
     super.onGateBlocked,
   });